introspector: Process a module's properties
We process the introspected data to determine a module's properties. We
get the information from the GObject property specs themselves, rather
than from doc comments, for now.

We only want to include properties that are usable from the app
description format. So we exclude non-writable properties in any case.

Ideally we'd want to exclude object-valued properties as well, but right
now the JSON format just contains either 'Number', 'Boolean', 'String',
or the GType name as the value type. The GType name could be an object,
boxed, or enum, so we can't just exclude everything that's not one of the
primitive types because we want to include enums. So instead we have a
somewhat clumsy whitelist of types.

We also want to exclude most of GTK's properties since they don't make
sense (or are too complicated) for the app description format. So we have
a blacklist of property origin classes, for widget classes whose
properties we don't want at all, and a whitelist of properties by class,
for widget classes from which we want only some of the properties.

https://phabricator.endlessm.com/T14916