introspector: Merge comments for the same symbol
The introspector constructs a doc comment for a property, populating its
fields with info from the introspected GParamSpec. However, there may
also already have been a doc comment about any given property in the
source code. In that case, we don't want to overwrite the existing doc
comment with the (usually more concise) information from the GParamSpec,
but we don't want to ignore it either. Instead we merge the two doc
comments.

https://phabricator.endlessm.com/T14916