formatter: Add symbol and formatting for slots
Slots are like properties, but instead of a value they have another
module or an array of modules. This adds slots to the documentation pages
and adds a rudimentary template for them. In our codebase currently we
don't have any documentation comments for them, so the info will be
pretty blank. It only shows whether the slot is array-valued (takes an
array of modules instead of a single module) or "multi" (the module
listed can be created more than once.)

https://phabricator.endlessm.com/T14916