formatter: Add symbol and formatting for references
References, instead of creating modules on demand like slots, point to
existing modules elsewhere in the module tree. This adds references to
the documentation pages and adds a rudimentary rendering template for
them. Like slots, we don't currently have any documentation comments for
references, so the info is pretty blank.

https://phabricator.endlessm.com/T14916