extension: Specify the path to 'introspect'
By default, we will look for the 'introspect' utility in the PATH.
However, we should be able to override it with a local path so that we
can build the eos-knowledge-lib documentation while uninstalled.

Similarly, we need a dependency on eos-knowledge-0-bin for the
'introspect' utility, but it can't be an absolute dependency, because
otherwise there would be a circular dependency.

https://phabricator.endlessm.com/T14916